requirements.txt
apache-airflow-providers-oracle
apache-airflow-providers-microsoft-mssql
apache-airflow-providers-amazon
apache-airflow-providers-snowflake
apache-airflow-providers-mysql
apache-airflow-providers-salesforce
airflow-provider-hightouch
snowflake-connector-python
snowflake-sqlalchemy
oracledb
numpy
scikit-learn
matplotlib
pandas
pyarrow
joblib
s3fs
boto3
salesforce-bulk
